requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-multipart==0.0.9
pandas>=2.0.0
openpyxl>=3.1.2
pyarrow>=14.0.1
numpy==1.26.4
pydantic==2.6.1
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
aiofiles==23.2.1
SQLAlchemy==2.0.27
alembic==1.13.1
psycopg2-binary==2.9.9  # Para PostgreSQL
redis==5.0.1  # Para cache (opcional)
streamlit==1.32.2
pillow==10.2.0
plotly==5.19.0
pytest==8.1.1
pytest-cov==4.1.0
black==24.2.0
isort==5.13.2
flake8==7.0.0
mypy==1.9.0 